[core] split Page behaviour into several traits
Page is getting huuuuuuuge, and it does several _almost_ completely-unrelated things
